worldmodels/requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.7.1
atari-py==0.1.7
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.159
botocore==1.12.159
Box2D==2.3.2
Box2D-kengz==2.3.3
certifi==2019.3.9
chardet==3.0.4
cloudpickle==1.2.1
cma==2.7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
docutils==0.14
EasyProcess==0.2.5
entrypoints==0.3
flake8==3.7.7
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.19.0
gym==0.9.4
h5py==2.9.0
hypothesis==4.21.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.0.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.2
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.3.1
PyOpenGL==3.1.0
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==4.4.0
python-dateutil==2.8.0
PyVirtualDisplay==0.2.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
tb-nightly==1.14.0a20190603
tensorflow==2.0.0b1
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019060501
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wrapt==1.11.2
